+++ src/log_formatter.erl
@@ -24,6 +24,7 @@
 -type config() :: #{debug => boolean(),
                     format => format(),
                     include_time => boolean(),
+                    color => boolean(),
                     atom() => any()}.
 
 -type msg() :: {io:format(), [term()]}